requirements.txt
openai>=1.0.0
python-dotenv
lxml
python-docx
streamlit
tqdm